Added User Model, Controller, Enpoints, CheckAuth middleware, and protected certain routes with checkAuth middleware.

+++ src/controllers/ExerciseController.ts
@@ -1,6 +1,7 @@
 import { Router, Request, Response, NextFunction } from 'express';
 import Exercise from '../models/Exercise';
 import { INewExercise } from '../models/interfaces/NewExerciseInterface';
+import { checkAuth, IRequestAuth } from '../middleware/check-auth';
 
 class ExerciseRouter {
 
@@ -62,12 +63,13 @@ class ExerciseRouter {
     });
   }
 
-  public createExercise(req: Request, res: Response, next: NextFunction): void {
+  public createExercise(req: IRequestAuth, res: Response, next: NextFunction): void {
     const newExercise: INewExercise = {
       name: req.body.name,
       sets: req.body.sets,
       reps: req.body.reps,
-      weight: req.body.weight
+      weight: req.body.weight,
+      creator: req.userData.userId
     };
 
     const exercise = new Exercise(newExercise);
@@ -91,9 +93,10 @@ class ExerciseRouter {
     });
   }
 
-  public updateExercise(req: Request, res: Response, next: NextFunction): void {
+  public updateExercise(req: IRequestAuth, res: Response, next: NextFunction): void {
     const id = req.params.id;
-    Exercise.findOneAndUpdate({ _id: id }, req.body)
+    const requestor = req.userData.userId
+    Exercise.findOneAndUpdate({ _id: id, creator: requestor }, req.body)
     .then((data) => {
       let status = res.statusCode;
       if (!data) {
@@ -120,9 +123,10 @@ class ExerciseRouter {
     });
   }
 
-  public deleteExercise(req: Request, res: Response, next: NextFunction): void {
+  public deleteExercise(req: IRequestAuth, res: Response, next: NextFunction): void {
     const id = req.params.id;
-    Exercise.findOneAndRemove({ _id: id })
+    const requestor = req.userData.userId
+    Exercise.findOneAndRemove({ _id: id, creator: requestor })
     .then((data) => {
       let status = res.statusCode;
       if (!data) {
@@ -151,10 +155,10 @@ class ExerciseRouter {
 
   public routes() {
     this.router.get('/', this.getExercises);
-    this.router.post('/', this.createExercise);
+    this.router.post('/', checkAuth, this.createExercise);
     this.router.get('/:id', this.getExercise);
-    this.router.put('/:id', this.updateExercise);
-    this.router.delete('/:id', this.deleteExercise);
+    this.router.put('/:id', checkAuth, this.updateExercise);
+    this.router.delete('/:id', checkAuth, this.deleteExercise);
   }
 
 }
